kivy_app/requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
docutils==0.22
filetype==1.2.0
fpdf==1.7.2
idna==3.10
Kivy==2.3.1
Kivy-Garden==0.1.5
pillow-avif-plugin==1.5.2
Pygments==2.19.2
requests==2.32.5
urllib3==2.5.0
